The authentication store file was moved from `frontend/src/js/auth.js` to `frontend/src/stores/auth.js` and the import path in `Login.vue` was updated accordingly.
